+++ Hbuilder_Example/nativeplugins/GCUniPlugin/ios/GC_UniPlugin_App.xcframework/ios-arm64/GC_UniPlugin_App.framework/Headers/GC_UniPlugin_App.h
@@ -1,0 +1,21 @@
+//
+//  GC_UniPlugin_App.h
+//  GC-UniPlugin-App
+//
+//  Created by hulilei on 2023/9/27.
+//
+
+#import <Foundation/Foundation.h>
+
+//! Project version number for GC_UniPlugin_App.
+FOUNDATION_EXPORT double GC_UniPlugin_AppVersionNumber;
+
+//! Project version string for GC_UniPlugin_App.
+FOUNDATION_EXPORT const unsigned char GC_UniPlugin_AppVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <GC_UniPlugin_App/PublicHeader.h>
+#import <GC_UniPlugin_App/FTMobileUniModule.h>
+#import <GC_UniPlugin_App/FTRUMModule.h>
+#import <GC_UniPlugin_App/FTTracerModule.h>
+#import <GC_UniPlugin_App/FTLogModule.h>
+