+++ UniPlugin-iOS/HBuilder-uniPluginDemo/GC-UniPlugin-App/GC-UniPlugin-App/Classes/FTUniPluginUtils.h
@@ -1,0 +1,16 @@
+//
+//  FTUniPluginUtils.h
+//  GC-UniPlugin-App
+//
+//  Created by hulilei on 2025/3/18.
+//
+
+#import <Foundation/Foundation.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface FTUniPluginUtils : NSObject
++ (BOOL)filterBlackResource:(NSURL *)url;
+@end
+
+NS_ASSUME_NONNULL_END